requirements_s3.txt
# Core Django
Django==4.2.10
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# Media handling - AWS S3 Storage
django-storages==1.14.2
boto3==1.34.0
Pillow==10.1.0

# Forms and UI
django-crispy-forms==2.1
crispy-bootstrap5==2023.10

# Maps & Location
geojson==2.5.0
django-leaflet==0.32.0

# Development
django-debug-toolbar==4.2.0
ipython==8.15.0
python-dotenv==1.1.1

# Production / Deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Security & Performance
django-extensions==4.1.0
django-filter==23.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.2

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Optional
django-bootstrap-icons==0.9.0
celery==5.3.6
redis==5.0.1
